backend/requirements.txt
asgiref==3.5.2
boto3==1.26.0
botocore==1.29.0
Django==4.0.5
django-rest-passwordreset==1.3.0
django-ses==3.3.0
django-storages==1.13.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
jmespath==1.0.1
mysqlclient==2.1.1
numpy==1.24.0
pandas==1.5.2
Pillow==9.3.0
PyJWT==2.4.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.12
